Prepare deprecation warnings for to-be-deleted functionality
But leave them commented out until we drop 0.4 support
